% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObliqueBART.R
\name{ObliqueBART}
\alias{ObliqueBART}
\title{Bayesian Additive Regression Trees with Oblique Splits}
\usage{
ObliqueBART(
  x,
  y,
  sparse = TRUE,
  ntrees = 10,
  node_min_size = 5,
  alpha = 0.95,
  beta = 2,
  nu = 3,
  lambda = 0.1,
  mu_mu = 0,
  sigma2 = 1,
  sigma2_mu = 1,
  k = 2,
  sigquant = 0.9,
  nburn = 1000,
  npost = 1000,
  nthin = 1,
  penalise_num_cov = FALSE,
  lambda_cov = 0.4,
  nu_cov = 2,
  coef_prior = "univariate_normal",
  coef_hyperprior = "none",
  coef_norm_hyperprior = "fixed",
  threshold_prior = "discrete_uniform",
  p_bar = ceiling(ncol(x)/3),
  norm_sigma_init = "Zellner",
  xi_prior = "Boojum",
  MH_propstep_xi = 0.15,
  norm_unit_sphere = FALSE,
  split_mix = FALSE,
  num_clust = 1
)
}
\arguments{
\item{x}{Training covariate matrix. Rows correspond to observations, columns correspond to covariates used in splitting rules.}

\item{y}{Training outcome vector.}

\item{coef_prior}{Prior distribution for splitting coefficients: "univariate_normal", "multivariate_normal", or "simplex"}

\item{coef_hyperprior}{Hyperprior for splitting coefficients.}

\item{p_bar}{Parameter if coef_hyperprior is univariate_normal_fixed_binomial. Mean number of covariates included in splitting rules.}
}
\value{
The following objects are returned:
\item{trees}{List (over MCMC iterations) of lists (over trees in a sum-of-trees model) of tree matrices.}
}
\description{
Bayesian Additive Regression Trees with Oblique Splits
}
